+++ app/settings-users/content-page.php
@@ -16,7 +16,7 @@ $resp->moduleName = $mmUsersT->NameModule;
 $resp->headTitle = $mmUsersT->NamePlural;
 $resp->headLogoutAuto = true;
 $resp->navInclude = true;
-$resp->contentHeader = $mmUsersT->FontAwesome . STR_NBSP . $mmUsersT->NamePlural . STR_NBSP;
+$resp->contentHeader = $mmUsersT->FontIcon . STR_NBSP . $mmUsersT->NamePlural . STR_NBSP;
 
 ///////////////////////////////////////////////////////////
 // Content Load Now or Later. Now is faster due to less 'round trips'. Later uses Ajax which is fast, but an extra 'round trip'.