+++ app/users-login/contentCard-user-login.php
@@ -40,7 +40,7 @@ JS;
 // Button
 $formButtonLoginTags = new \xan\tags( [ ELE_CLASS_BUTTON_RG_PRIMARY ], [], [ "onclick='" . $formButtonLoginOnClick . "'" ] );
 $formButtonLoginSpinner = '<span id="formButtonLoginSpinner" style="display: none;">' . STR_NBSP . FI_SPINNER . '</span>';
-$buttonEle = new \xan\eleButton( $mmUsersLogin->FontAwesome . STR_NBSP . $mmUsersLogin->NameModule . $formButtonLoginSpinner, 'formButtonLogin', '', $formButtonLoginTags );
+$buttonEle = new \xan\eleButton( $mmUsersLogin->FontIcon . STR_NBSP . $mmUsersLogin->NameModule . $formButtonLoginSpinner, 'formButtonLogin', '', $formButtonLoginTags );
 $tableLogin->cellSet( ++$tableRowIndex, 0, $tagsCellLeftMiddle, $buttonEle->render() );
 
 // Message
@@ -76,7 +76,7 @@ JS;
 // Button
 $formButtonCodeTags = new \xan\tags( [ ELE_CLASS_BUTTON_RG_PRIMARY ], [], [ "onclick='" . $formButtonCodeOnClick . "'" ] );
 $formButtonCodeSpinner = '<span id="formButtonCodeSpinner" style="display: none;">' . STR_NBSP . FI_SPINNER . '</span>';
-$buttonEle = new \xan\eleButton( $mmUsersLogin->FontAwesome . STR_NBSP . 'Enter Code to Login' . $formButtonCodeSpinner, 'formButtonCode', '', $formButtonCodeTags );
+$buttonEle = new \xan\eleButton( $mmUsersLogin->FontIcon . STR_NBSP . 'Enter Code to Login' . $formButtonCodeSpinner, 'formButtonCode', '', $formButtonCodeTags );
 $tableCode->cellSet( ++$tableRowIndex, 0, $tagsCellLeftMiddle, $buttonEle->render() );
 
 // Message